setup.py now properly includes header files

+++ py/setup.py
@@ -226,14 +226,24 @@ if CXX:
 #
 
 version = None
-if os.path.exists(os.path.join(thisdir, '..', 'VERSION')):
-    with open('../VERSION') as f:
-        version = ensure_str(f.read()).strip()
 try:
     version = ensure_str(subprocess.check_output([vv.get('GIT'), 'describe', '--tags', 'HEAD'])).strip()
-except Exception as e:
-    print("ERROR: Can't retrieve the current code version.")
-    raise
+except Exception:
+    pass
+if not version:
+    if os.path.exists(os.path.join(thisdir, '..', 'VERSION')):
+        with open(os.path.join(thisdir, '..', 'VERSION')) as f:
+            version = ensure_str(f.read()).strip()
+    elif os.path.exists(os.path.join(thisdir, 'VERSION')):
+        with open(os.path.join(thisdir, 'VERSION')) as f:
+            version = ensure_str(f.read()).strip()
+if not version:
+    raise RuntimeError("Cannot determine Tomographer version (no git info and no VERSION file)")
+
+# create VERSION file here -- for sdist
+with open(os.path.join(thisdir, 'VERSION'), 'w') as f:
+    f.write(version + "\n")
+
 
 # major/minor sections of version
 m = re.match(r'^v(?P<major>\d+)\.(?P<minor>\d+)', version)
@@ -430,23 +440,62 @@ class BuildExt(build_ext):
 #
 # Copy all tomographer headers inside tomographer.include as package data
 #
-shutil.rmtree(os.path.join(thisdir, 'tomographer', 'include', 'tomographer'))
-shutil.copytree(os.path.join(thisdir, '..', 'tomographer'),
-                os.path.join(thisdir, 'tomographer', 'include', 'tomographer'),
-                ignore=lambda d, files: [f for f in files if not f.endswith('.h')])
-# create tomographer_version.h
-tomographer_version_t = None
-with open(os.path.join(thisdir, '..', 'tomographer', 'tomographer_version.h.in')) as f:
-    tomographer_version_t = f.read()
+src_tomographer_include = os.path.join(thisdir, '..', 'tomographer')
+target_tomographer_include = os.path.join(thisdir, 'tomographer', 'include')
+if os.path.exists(src_tomographer_include):
+    if os.path.exists(os.path.join(target_tomographer_include, 'tomographer')):
+        shutil.rmtree(os.path.join(target_tomographer_include, 'tomographer'))
+    shutil.copytree(src_tomographer_include,
+                    os.path.join(target_tomographer_include, 'tomographer'),
+                    ignore=lambda d, files: [f for f in files
+                                             if not os.path.isdir(os.path.join(d,f)) and not f.endswith('.h')])
+elif not os.path.exists(target_tomographer_include):
+    raise RuntimeError("Can't import tomographer headers in package, source doesn't exist!")
 
-repvars = {'TOMOGRAPHER_VERSION': version,
-           'TOMOGRAPHER_VERSION_MAJ': str(version_maj),
-           'TOMOGRAPHER_VERSION_MIN': str(version_min) }
-tomographer_version_f = re.sub(r'@(?P<var>[A-Za-z_0-9]+)@', lambda m: repvars[m.group('var')], tomographer_version_t)
+# create tomographer_version.h
+tomographer_version_h_content = """\
+/* This file is part of the Tomographer project, which is distributed under the
+ * terms of the MIT license.
+ *
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016 ETH Zurich, Institute for Theoretical Physics, Philippe Faist
+ * Copyright (c) 2017 Caltech, Institute for Quantum Information and Matter, Philippe Faist
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#ifndef zzzTOMOGRAPHER_VERSION_H_
+#define zzzTOMOGRAPHER_VERSION_H_
+
+#define TOMOGRAPHER_VERSION "{version}"
+#define TOMOGRAPHER_VERSION_MAJ {version_maj}
+#define TOMOGRAPHER_VERSION_MIN {version_min}
+
+#endif
+""".format(version=version,version_maj=version_maj,version_min=version_min)
 with open(os.path.join(thisdir, 'tomographer', 'include', 'tomographer', 'tomographer_version.h'), 'w') as f:
-    f.write(tomographer_version_f)
+    f.write(tomographer_version_h_content)
+
 # include license file in this directory
-shutil.copy2(os.path.join(thisdir, '..', 'LICENSE.txt'), thisdir)
+if os.path.exists(os.path.join(thisdir, '..', 'LICENSE.txt')):
+    shutil.copy2(os.path.join(thisdir, '..', 'LICENSE.txt'), thisdir)
 
 
 #
@@ -476,8 +525,7 @@ setup(name="tomographer",
                   numpy.get_include(),
                   vv.get("Boost_INCLUDE_DIR"),
                   vv.get("EIGEN3_INCLUDE_DIR"),
-                  os.path.join(thisdir, ".."), # tomographer
-                  os.path.join(thisdir, "tomographer/include"), # tomographerpy
+                  os.path.join(thisdir, "tomographer/include"), # tomographer & tomographerpy
               ],
               depends=headers,
               language='c++',
